+++ applications/display/lib/bootstrapManifest.js
@@ -16,99 +16,76 @@
 
 var path = require("path");
 var Thywill = require("thywill");
-
 var Resource = Thywill.getBaseClass("Resource");
-var encoding = "utf8";
 
 var manifest = {
   // Add Modernizr, which has to come first in the Javascript.
   "../../../thirdParty/modernizr/modernizr.2.6.1.min.js": {
     clientPath: "/display/js/modernizr.min.js",
-    encoding: encoding,
-    minified: true,
-    type: Resource.TYPES.JAVASCRIPT,
     weight: -30
   },
   // Add jQuery as a resource, setting it a lower weight than the default
   // Thwyill code - having it come first is fairly necessary if you want
   // things to work rather than explode.
-  "../../../thirdParty/jquery/jquery.1.7.2.min.js": {
+  "../../../thirdParty/jquery/jquery.1.9.1.min.js": {
     clientPath: "/display/js/jquery.min.js",
-    encoding: encoding,
-    minified: true,
-    type: Resource.TYPES.JAVASCRIPT,
     weight: -20
   },
   // Add the plugins.js code from HTML5 Boilerplate.
   "../../../thirdParty/html5boilerplate/plugins.js": {
     clientPath: "/display/js/plugins.js",
-    encoding: encoding,
-    minified: false,
-    type: Resource.TYPES.JAVASCRIPT,
     weight: -10
   },
   // Add Handlebars.js.
   "../../../thirdParty/handlebars.js/handlebars.1.0.0.rc.1.js": {
     clientPath: "/display/js/handlebars.js",
-    encoding: encoding,
-    minified: false,
-    type: Resource.TYPES.JAVASCRIPT,
     weight: 10
   },
   // Add dateformat.js.
   "../../../thirdParty/dateformat/dateformat.1.2.3.js": {
     clientPath: "/display/js/dateformat.js",
-    encoding: encoding,
-    minified: false,
-    type: Resource.TYPES.JAVASCRIPT,
     weight: 20
   },
   // Add HTML5 Boilerplate CSS.
   "../../../thirdParty/html5boilerplate/html5boilerplate.css": {
     clientPath: "/display/css/html5boilerplate.css",
-    encoding: encoding,
-    minified: false,
-    type: Resource.TYPES.CSS,
     weight: 0
   },
   // Add the Display client CSS.
   "../client/css/displayClient.css": {
     clientPath: "/display/css/client.css",
-    encoding: encoding,
-    minified: false,
-    type: Resource.TYPES.CSS,
     weight: 10
   },
   // Add the Display client UI templates. Note that these won't be loaded over
   // HTTP, but rather included into the application main page.
   "../client/template/ui.tpl": {
     clientPath: "/display/tpl/ui.tpl",
-    encoding: encoding,
     id: "display-template-ui",
     type: Resource.TYPES.TEMPLATE,
     weight: 0
   },
   "../client/template/text.tpl": {
     clientPath: "/display/tpl/text.tpl",
-    encoding: encoding,
     id: "display-template-text",
     type: Resource.TYPES.TEMPLATE,
     weight: 0
   },
   "../client/template/connection.tpl": {
     clientPath: "/display/tpl/connection.tpl",
-    encoding: encoding,
     id: "display-template-connection",
     type: Resource.TYPES.TEMPLATE,
     weight: 0
   }
 };
 
-// Convert all the relative paths to absolute paths.
+// Convert all the relative paths to absolute paths, and set the encoding
+// while we're about it.
 var absolutePath, absoluteManifest = {};
+var encoding = "utf8";
 for (var originFilePath in manifest) {
   absolutePath = path.resolve(__dirname, originFilePath);
   absoluteManifest[absolutePath] = manifest[originFilePath];
+  absoluteManifest[absolutePath].encoding = encoding;
 }
 
 //-----------------------------------------------------------