+++ applications/display/lib/service.js
@@ -137,7 +137,6 @@ exports.start = function (clusterMemberId) {
     },
     cacheSize: 100,
     redisPrefix: "thywill:display:resource:",
-    // This will be set in the start script.
     redisClient: redisClients.resourceManager
   };