+++ applications/draw/lib/draw.js
@@ -64,9 +64,7 @@ p._defineBootstrapResources = function (callback) {
       var resource = self.thywill.resourceManager.createResource(data, {
         clientPath: "/draw/js/drawClient.js",
         encoding: encoding,
-        minified: false,
         originFilePath: originFilePath,
-        type: self.thywill.resourceManager.types.JAVASCRIPT,
         weight: 50
       });
       self.storeBootstrapResource(resource, asyncCallback);