+++ applications/echo/service/startEcho.js
@@ -28,8 +28,8 @@ config.clientInterface.socketConfig.global.resource = "/echo/socket.io";
 config.clientInterface.socketConfig.global.store = new MemoryStore();
 
 // Resource minification settings.
-config.clientInterface.minifyCss = true;
-config.clientInterface.minifyJavascript = true;
+config.clientInterface.minifyCss = false;
+config.clientInterface.minifyJavascript = false;
 
 // Set an appropriate log level for an example application.
 config.log.level = "debug";